pyproject.toml
[tool.poetry]
name = "podcast-generator"
version = "0.1.0"
description = ""
authors = ["Lucasgyn94 <needslucas944@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.35.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
